Seperate utf8 module into directory
