+++ noline/src/editor/mod.rs
@@ -1,5 +1,5 @@
+mod async_editor;
 pub mod builder;
 mod sync_editor;
-mod async_editor;
 
 pub use builder::*;
