+++ noline/src/utf8/mod.rs
@@ -1,0 +1,9 @@
+mod byte;
+mod char;
+mod decoder;
+
+pub use char::*;
+pub use decoder::*;
+
+#[cfg(test)]
+pub(crate) mod tests;
